tests: refactor code tests and add more scenarios
