+++ src/game/gameTimer.js
@@ -2,6 +2,10 @@
 let gameTime;
 
 function setTimer(time = 10, finishGame) {
+  if (time < 1 || time > 60 || typeof time !== "number") {
+    const err = new Error("Invalid value for game time.");
+    throw err;
+  }
   const timer = document.querySelector(".timer");
   let totalSeconds = time * 60;
   let minutes = totalSeconds / 60;
